feat: Web Worker for PDF gen
